refactor(canvas): break import-worker into focused modules
extract the 1488-line monolithic import worker into five
focused modules with clear responsibilities:
- async-limiter.js: concurrency primitives (createAsyncLimiter, pooled)
- import-metrics.js: timing, logging, env-parsing helpers
- import-discovery.js: two-phase discovery and legacy course processing
- import-extraction.js: file download, dedup, import record management
- import-embedding.js: RAG pipeline (content extraction + embedding)

the worker file becomes a ~120-line orchestrator that imports
and re-exports from the modules. all functionality, error paths,
retry logic, and message interfaces are preserved.
